chore(ui): rename dumb render components to Base...
