+++ ui/src/components/BaseMewList.vue
@@ -2,7 +2,7 @@
   <h6 v-if="title" class="q-mb-md">{{ title }}</h6>
 
   <MewListSkeleton v-if="isLoading || !feedMews" />
-  <EmptyMewsFeed v-else-if="feedMews && !isLoading && feedMews.length === 0" />
+  <BaseEmptyMewsFeed v-else-if="feedMews && !isLoading && feedMews.length === 0" />
   <QList v-else bordered separator>
     <BaseMewListItem
       v-for="(mew, index) of feedMews"
@@ -18,7 +18,7 @@ import { QList } from "quasar";
 import BaseMewListItem from "@/components/BaseMewListItem.vue";
 import MewListSkeleton from "@/components/MewListSkeleton.vue";
 import { FeedMew } from "@/types/types";
-import EmptyMewsFeed from "./EmptyMewsFeed.vue";
+import BaseEmptyMewsFeed from "./BaseEmptyMewsFeed.vue";
 
 withDefaults(
   defineProps<{
