+++ ui/src/components/BaseMewListItem.vue
@@ -228,7 +228,7 @@
       v-model="showCreateMewmewDialog"
       @profile-created="createMewmew"
     />
-    <ConfirmDialog
+    <BaseConfirmDialog
       v-model="showConfirmDeleteDialog"
       title="Delete Mew"
       confirm-text="Delete"
@@ -239,7 +239,7 @@
         Note that other peers may still have copies of the data, and you can't
         force them to delete it.
       </p>
-    </ConfirmDialog>
+    </BaseConfirmDialog>
   </QItem>
 </template>
 
@@ -261,7 +261,7 @@ import { useRouter } from "vue-router";
 import { AppAgentClient } from "@holochain/client";
 import BaseTimestamp from "./BaseTimestamp.vue";
 import CreateProfileIfNotFoundDialog from "@/components/CreateProfileIfNotFoundDialog.vue";
-import ConfirmDialog from "@/components/ConfirmDialog.vue";
+import BaseConfirmDialog from "@/components/BaseConfirmDialog.vue";
 import { showMessage } from "@/utils/notification";
 import dayjs from "dayjs";
 import BaseEmbedMew from "@/components/BaseEmbedMew.vue";
