+++ ui/src/components/BaseProfileSkeleton.vue
